requirements.txt
websockets==11.0.3
aiohttp==3.7.4
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
openai==1.12.0
certifi==2024.2.2
